Fix apn display issue on multisim
Subscription id is passed with intent to launch ApnEditor with an
extra string "subscription", which is changed to "sub_id". With this
change, sub id cannot be retrieved, and Dds sub id is considered for
apn display. Hence Dds sub apn's are displayed for both subs.
Fix is to restore extra string in intent to "subscription".

Change-Id: Icb717d376c1670e5f9afc742bcf8f8a09e03adc3
CRs-Fixed: 808987