IMS: Add to unknown connection if connection is alive.
When a CS call is disconnected from remote side, the
disconnected connection is not cleared till next call
state change. This causes an issue for phantom calls on
IMS after a CS call end. The CS connection in disconncted
state gets picked up wrongly as the unknown connection
for phantom IMS call and results in wrong call states for
the IMS call.
Fix this by adding foreground connection to the unknown
connection list if it's not a disconnected or disconnecting
connection.

Change-Id: I95b5d46221885b8d8d6acf855fde51bba92bbcdf
CRs-Fixed: 841686